Change 252 by jain156@akshayj-lt1 on 2017/05/29 10:51:32
Checking in the data footprint stats as reported in the micro paper.

The implementation uses gpgpu pointer, due to which I don't feel fine pushing to mainline.
May be I can do the stats check and increment in l2cache.cc through a public function
to avoid exposing gpgpu pointer to public