gpgpu-sim/shader: Calculate max pipeline depth when int is executed by SF.
Fixes a problem exposed (but not introduced) by:
  d212d7e: take account of shfl latency

Under the SM2_GTX480 configuration, a shfl instruction now results in an out-of-
bounds write into a non-allocated pipeline slot, causing "random" crashes.

While at it, tidy up a bit. There doesn't seem to be a strongly enforced coding
convention, but three tabs for one level of identation seems a tad excessive
for any style.

Signed-off-by: Roy Spliet <rs855@cam.ac.uk>